% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shlib.R
\name{shlib}
\alias{shlib}
\title{Compile shared library}
\usage{
shlib(filenames, verbose = TRUE, output = NULL, clean = FALSE,
  preclean = FALSE, dry_run = FALSE, debug = FALSE)
}
\arguments{
\item{filenames}{Vector of filenames}

\item{verbose}{Be verbose (print compiler output to screen)}

\item{output}{Name to use for the file output (passed to
\code{SHLIB} with \code{--output} -- by default the name of the
first file is used).}

\item{clean}{Clean intermediate targets (passed to
\code{SHLIB} as \code{--clean})}

\item{preclean}{Clean intermediate targets before running (passed to
\code{SHLIB} as \code{--clean})}

\item{dry_run}{Don't run anything, but print commands (passed to
\code{SHLIB} as \code{--dry-run})}

\item{debug}{Create a debug dll (windows only; passed to
\code{SHLIB} as \code{--debug})}
}
\description{
Compile a shared library using \code{R CMD SHLIB}.
}
\author{
Rich FitzJohn
}

